Create RegistrationService for AB#16811
