+++ Apps/GatewayApi/src/Controllers/UserProfileControllerV2.cs
@@ -49,6 +49,7 @@ namespace HealthGateway.GatewayApi.Controllers
         private readonly IUserPreferenceServiceV2 userPreferenceService;
         private readonly ILegalAgreementServiceV2 legalAgreementService;
         private readonly IUserValidationService userValidationService;
+        private readonly IRegistrationService registrationService;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="UserProfileControllerV2"/> class.
@@ -60,6 +61,7 @@ namespace HealthGateway.GatewayApi.Controllers
         /// <param name="userPreferenceService">The injected user preference service.</param>
         /// <param name="legalAgreementService">The injected legal agreement service.</param>
         /// <param name="userValidationService">The injected user validation service.</param>
+        /// <param name="registrationService">The injected registration service.</param>
         public UserProfileControllerV2(
             IUserProfileServiceV2 userProfileService,
             IHttpContextAccessor httpContextAccessor,
@@ -67,7 +69,8 @@ namespace HealthGateway.GatewayApi.Controllers
             IUserSmsServiceV2 userSmsService,
             IUserPreferenceServiceV2 userPreferenceService,
             ILegalAgreementServiceV2 legalAgreementService,
-            IUserValidationService userValidationService)
+            IUserValidationService userValidationService,
+            IRegistrationService registrationService)
         {
             this.userProfileService = userProfileService;
             this.httpContextAccessor = httpContextAccessor;
@@ -76,6 +79,7 @@ namespace HealthGateway.GatewayApi.Controllers
             this.userPreferenceService = userPreferenceService;
             this.legalAgreementService = legalAgreementService;
             this.userValidationService = userValidationService;
+            this.registrationService = registrationService;
         }
 
         /// <summary>
@@ -116,7 +120,7 @@ namespace HealthGateway.GatewayApi.Controllers
             ClaimsPrincipal user = this.httpContextAccessor.HttpContext!.User;
             DateTime jwtAuthTime = ClaimsPrincipalReader.GetAuthDateTime(user);
             string? jwtEmailAddress = user.FindFirstValue(ClaimTypes.Email);
-            return await this.userProfileService.CreateUserProfileAsync(createUserRequest, jwtAuthTime, jwtEmailAddress, ct);
+            return await this.registrationService.CreateUserProfileAsync(createUserRequest, jwtAuthTime, jwtEmailAddress, ct);
         }
 
         /// <summary>
