+++ Apps/GatewayApi/src/Services/IRegistrationService.cs
@@ -1,0 +1,38 @@
+// -------------------------------------------------------------------------
+//  Copyright © 2019 Province of British Columbia
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+// -------------------------------------------------------------------------
+namespace HealthGateway.GatewayApi.Services
+{
+    using System;
+    using System.Threading;
+    using System.Threading.Tasks;
+    using HealthGateway.GatewayApi.Models;
+
+    /// <summary>
+    /// The registration service.
+    /// </summary>
+    public interface IRegistrationService
+    {
+        /// <summary>
+        /// Creates a user profile.
+        /// </summary>
+        /// <param name="createProfileRequest">The request to create a user profile model.</param>
+        /// <param name="jwtAuthTime">The authenticated login time from the JWT.</param>
+        /// <param name="jwtEmailAddress">The email address contained in the JWT.</param>
+        /// <param name="ct"><see cref="CancellationToken"/> to manage the async request.</param>
+        /// <returns>The user profile.</returns>
+        Task<UserProfileModel> CreateUserProfileAsync(CreateUserRequest createProfileRequest, DateTime jwtAuthTime, string? jwtEmailAddress, CancellationToken ct = default);
+    }
+}
